split wiki command
